refactor: refactor library
