+++ package.json
@@ -26,23 +26,24 @@
     "@discordjs/opus": "^0.4.0",
     "axios": "^0.21.1",
     "common-tags": "^1.8.0",
-    "discord-ytdl-core": "^5.0.0",
+    "discord-ytdl-core": "^5.0.1",
     "discord.js": "^12.5.1",
-    "discord.js-commando": "https://github.com/discordjs/Commando.git#commit=9c2e5405fdb9a1006e6eb7554e18d0861a4c857f",
+    "discord.js-commando": "^0.12.2",
     "dotenv": "^8.2.0",
     "ffmpeg-static": "^4.2.7",
     "merge-img": "^2.1.3",
-    "mongodb": "^3.6.3",
-    "mongoose": "^5.11.13",
+    "mongodb": "^3.6.4",
+    "mongoose": "^5.11.18",
     "node-fetch": "^2.6.1",
-    "node-gtrans": "^1.5.1",
+    "node-gtrans": "^1.6.0",
     "winston": "^3.3.3",
-    "yt-search": "^2.5.2",
-    "ytdl-core": "^4.4.5",
-    "ytpl": "^2.0.4"
+    "yt-search": "^2.7.3",
+    "ytdl-core": "^4.5.0",
+    "ytpl": "^2.0.5"
   },
   "devDependencies": {
-    "eslint": "^7.18.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.20.0",
     "eslint-plugin-import": "^2.22.1"
   }
 }
