+++ src/Modules/administration/setgoodbye.js
@@ -1,6 +1,6 @@
 const { stripIndents } = require('common-tags');
 const Command = require('../../structures/Command.js');
-const { guildSettingsSchema } = require('../../library/Database/schema.js');
+const { guildSettingsSchema } = require('../../util/schema.js');
 
 
 module.exports = class GoodbyeCommand extends Command {
