+++ src/events/message.js
@@ -1,5 +1,5 @@
 const { client } = require('../bot.js');
-const { guildSettingsSchema, userDataSchema } = require('../library/Database/schema.js');
+const { guildSettingsSchema, userDataSchema } = require('../util/schema.js');
 
 // event on message
 client.on('message', async msg => {
